Organize project into packages
- Move classes into separate packages:
  - `duke.command`
  - `duke.exception`
  - `duke.task`
  - `duke.ui`
  - `duke.storage`

- Update import statements in affected classes
